refactor: 레이아웃 구조 개선 및 common 컴포넌트 이동
