--- frontend/src/pages/layouts/index.ts
@@ -1,1 +1,0 @@
-export * from './RootLayout'
